Rotation and moving of furniture.
Objects can be moved along x and z axis using the wasd keyboard keys. To
rotate use q and e keys. The object is rotating along its center by
calculating the minimum and maximum of its verteces in order to
determine the measure of the object. Then to rotate it along its center,
firts the apply a translate over the viewMatrix on x and z using those
values /2, then the viewMatrix is multiplied by the rotationMatrix, and
then the object is translated back.