Object movement and grid plane.
Skybox no more used, added a grid plane created with Blender as a
working space area. Objects now can be moved along all the axes, rotated, scaled. A partial collision detection works. Problems occure when object is scaled or rotated. Moreover when an object is moved, scaled and rotated, rotation is no more along the center of the object. Collision not working when rotating. To be implemented soon.