OBB collision implementation.
The collision detection is an obb using the SAT theorem. Now collision
is detected also when rotation is applied to an object, not only while
moving along axes. The AABB technique is replaced by the OBB. AABB is
the axis aligned bounding box while the OBB is the oriented bounding
box. The Separating Axis Theorem (SAT) says that two convex objects
do not overlap if there exists a line (called axis) onto which the
two objects' projections do not overlap.