Origin moved to the real object origin.
The problem with the scaling and collision not working, now is adjusted.
There was a problem in the representation of the origin of the object,
before it was the left side point of the object now it is the real
center of the object. Objects can be proportionally scaled without
problems. A feature for enabling and disabling collision is added, in
order to simplify the insertion of the furniture preventing the
collision while inserting new furniture and so creating a stall.