Code refactoring.
New smaller files with code functions are now present. The app core with
the main is in app.js, other stuff moved in help.js, collision code is
in the corresponding file, event handlers are in the handler.js, jquery
code is in menu.js, model furnitures are in model.js.