New todo added.
Ray casting should be better for selection of an object. For the moment,
the pick shader is used, not really optimized. What happens is that in
one frame, the pick shader color is drawn and after that in the same
frame the textures are also applied. <- not really optimized so